refactor: modularize parser components for better maintainability
Modularised the parser module into seperate modules for different file elements.
